fix: move things around, update copy, validation
